Clean up formatting, remove tabs, and to Jamfile.v2.
